chore(swingset): add tools to classify mainnet runs
Feed a slogfile into stdin of `classify-runs.py`, and it will emit one
line per run, describing why the run was performed (price-feed,
psm-sell, etc, known as the "type" of the run), along with the
b${blockNum}-r${runNum} identifier, the number of computrons consumed,
and the number of wallclock seconds consumed.

Feed all those output lines into the stdin of `classify-types.py`, and
it will emit a table of statistics about those types, like the median
wallclock time consumed.